refactor(policy): proxy-based enforcement — identical behavior on Node and Bare
Replaces the AsyncLocalStorage-based wrapper (and its no-op Bare fallback)
with a single Proxy-based mechanism that works identically on every
runtime that supports ES Proxy.

How it works
- `_applyPolicies` returns a Proxy wrapping the underlying account
  instead of mutating it. WDK manager returns this proxy from
  `getAccount` / `getAccountByPath`.
- The proxy's `get` trap exposes enforced versions of write methods.
- For every non-enforced property, the trap returns the underlying value;
  if it's a function, it's bound to the underlying account.
- Internal SDK code that calls `this.someMethod()` resolves on the
  underlying account, not the proxy — bypassing enforcement naturally.
- Protocol packages that hold a direct reference to the account
  (constructed before the proxy exists) likewise bypass enforcement.
- Only access through the proxy goes through policy evaluation.

Why this is better than what we had
- Identical code path on Node and Bare. No `AsyncLocalStorage`, no
  `node:async_hooks` dependency, no runtime detection, no
  Promise-prototype patching, no fallback.
- Eliminates the divergence we shipped previously (Node had
  nested-call escape, Bare didn't). Both now work the same way.
- Doesn't depend on V8 internals or TC39 AsyncContext landing.
- The original account stays clean — easier to reason about and
  matches the "policy proxy is a view, not a mutation" mental model
  the dev portal will need in later phases.

Files
- Added: `src/policy/policy-account-proxy.js`
- Removed: `src/policy/policy-account-wrapper.js`,
  `src/policy/policy-context-store.js`,
  `tests/wdk-manager-policy-bare.test.js` (the Bare-divergence tests
  no longer apply — coverage for nested escape is in the main test
  file and works the same on every runtime).
- Updated: `src/policy/policy-engine.js` (calls the new factory),
  `src/wdk-manager.js` (returns the proxy), README + AGENTS.md
  (drop the "Bare-specific limitation" section).

Tests + types + lint + smoke all clean: 101/101 tests, 51/51 smoke.
