+++ AGENTS.md
@@ -53,7 +53,7 @@ Source code must be strictly typed using JSDoc comments to support the `build:ty
 - The engine wraps account write methods and protocol getters at `getAccount` time. Wrapping is dynamic — only methods named in registered rules are wrapped, and only when at least one policy applies.
 - Two scopes only: `project` (with optional wallet restriction supplied via the first argument to `registerPolicy`) and `account` (with required wallet binding and a per-account `accounts` list of paths and/or integer indexes). The "wallet" argument is the same string passed to `registerWallet` — an opaque consumer-chosen key, not a blockchain identifier. DENY wins across scopes; account-scope `ALLOW` rules can short-circuit project-scope DENYs via `override_broader_scope: true`.
 - Index-form `accounts` entries match the index passed to `wdk.getAccount(wallet, index)`. They do not match accounts retrieved via `getAccountByPath` because the wallet manager has no synchronous path → index resolver. Path-form entries match either retrieval style.
-- The "in policy context" marker uses `AsyncLocalStorage` on Node (per-async-chain): concurrent calls on the same account each evaluate independently, while nested calls within one chain skip re-evaluation. The store is selected at first use via `src/policy/policy-context-store.js`. On runtimes without `AsyncLocalStorage` (Bare), the store falls back to a no-op — the engine still works for every other policy behavior, but nested-call escape is disabled (every wrapped call evaluates independently, including nested). Modern V8 inlined `await` to skip user-visible `Promise.prototype.then` years ago, so the Zone.js technique no longer works in pure user-space; we wait for TC39 AsyncContext + V8 + Bare to land before the Bare path matches Node. `node:async_hooks` is loaded lazily so importing WDK on Bare does not crash even if no policy is registered.
+- Nested-call escape works via a `Proxy` (see `src/policy/policy-account-proxy.js`), not an async-context marker. The engine returns a Proxy that exposes enforced versions of write methods; the proxy's `get` trap binds non-enforced methods to the underlying account, so internal `this.method()` calls and protocol code that holds a direct reference to the account naturally bypass enforcement. The original account is never mutated. The same code path runs on every JavaScript runtime that supports `Proxy` — no `AsyncLocalStorage`, no runtime detection, no Bare-specific fallback.
 - Each condition is raced against `conditionTimeoutMs` (default 30s, settable via `RegisterPolicyOptions`). A condition that throws or times out is fail-closed for DENY rules (treated as matched → block) and fail-open-as-no-match for ALLOW rules.
 - Conditions are user-supplied functions in Phase 1. The engine accepts `state` and `onSuccess` rule fields for Phase 2 (engine-managed state + post-execution hooks) but ignores them at runtime; `state` is deep-cloned at registration so callers can't mutate engine state post-registration.
 - Tests live in `tests/wdk-manager-policy.test.js` and exercise the engine exclusively through the public WDK API.
