+++ README.md
@@ -100,9 +100,7 @@ const result = await account.simulate.sendTransaction({ to: '0x…', value: 1n }
 
 Policies have two scopes — `project` and `account`. A project-scope policy applies globally by default, or only to the wallets you bind it to (`wdk.registerPolicy('ethereum', policy)` or `wdk.registerPolicy(['ethereum', 'ton'], policy)`). The first argument is a wallet identifier — the same string passed to `registerWallet`. It might be a chain name like `"ethereum"`, but it could equally be `"treasury-cold"` or any label the consumer chose; the engine treats it as an opaque key. An account-scope policy targets specific accounts within a wallet, identified by either derivation path (`accounts: ["0'/0/0"]`) or integer index (`accounts: [0, 1]`) — index entries match accounts retrieved via `wdk.getAccount(wallet, index)`; path entries match either retrieval style. Evaluation is narrowest-first with `DENY` winning across scopes. Account-scope `ALLOW` rules can opt into `override_broader_scope: true` to short-circuit broader policies for explicit exceptions (e.g., treasury accounts). Conditions can be sync or async and may carry user-owned state via closures. Templates (`@tetherto/wdk-policy-templates`) and a portal UI for editing policies are coming in later phases.
 
-### Runtime support
-
-The policy engine works on both Node.js and the Bare runtime. Behavior differs in one place: **nested-call escape**. On Node, when a wrapped method (e.g. `bridge`) internally invokes another wrapped method (e.g. `sendTransaction`), the inner call inherits the outer call's policy context and is not re-evaluated. On Bare, every wrapped call evaluates independently — including nested ones — because Bare does not yet expose `AsyncLocalStorage` (the Holepunch team is waiting on TC39 AsyncContext to reach Stage 3 + V8 implementation). This matters when, for example, a project policy denies `sendTransaction` while allowing `bridge`: on Node the `bridge` call succeeds, on Bare the inner `sendTransaction` re-trips the DENY and the bridge fails. Write your Bare-side policies so high-level operations (`bridge`, `swap`, etc.) and the underlying `sendTransaction` are both compatible with each other, or stage your DENYs at the highest-level operation only.
+The engine wraps accounts through an ES `Proxy` so internal SDK code that uses `this.method()` naturally bypasses enforcement — nested-call escape (e.g. `bridge` internally calling `sendTransaction`) works without any async-context tracking. The same code path runs on every JavaScript runtime that supports `Proxy`, including Bare.
 
 ## Compatibility
